rename shared -> application
