+++ InnoShop.Application/Shared/Auth/BaseResult.cs
