+++ InnoShop.Application/Shared/Auth/LoginResult.cs
