+++ InnoShop.Application/Shared/Auth/UserInfoDto.cs
