+++ InnoShop.Application/Shared/Misc.cs
