+++ InnoShop.Domain/Entities/ShopUser.cs
@@ -1,4 +1,5 @@
 ﻿namespace InnoShop.Domain;
+
 using Microsoft.AspNetCore.Identity;
 
 public class ShopUser : IdentityUser { }
