+++ InnoShop.Tests/GlobalUsings.cs
@@ -1,1 +1,4 @@
 global using NUnit.Framework;
+global using Microsoft.AspNetCore.Mvc.Testing;
+global using Microsoft.EntityFrameworkCore;
+global using Microsoft.Extensions.DependencyInjection;
