+++ InnoShop.Tests/TestUserAPI/TestSetupApi.cs
@@ -1,16 +1,9 @@
 using InnoShop.Application.Shared.Auth;
 using InnoShop.Infrastructure.UserManagerAPI;
-using InnoShop.Infrastructure.UserManagerAPI.Data;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.DependencyInjection;
 using System.Net.Http.Json;
-using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.AspNetCore.Hosting;
-using System.Data.Common;
 using Newtonsoft.Json;
 using System.Net.Http.Headers;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
-using InnoShop.Domain.Services;
 using System.Net.Mail;
 
 namespace InnoShop.Tests.TestUserAPI;
@@ -18,7 +11,6 @@ namespace InnoShop.Tests.TestUserAPI;
 public class TestSetupApi {
     WebApplicationFactory<Program> webApp;
     protected HttpClient client;
-
 
     [SetUp]
     public virtual Task SetupAsync() {
@@ -85,55 +77,5 @@ public class TestSetupApi {
             return result;
 
         throw new JsonSerializationException($"Cannot convert {jsonString} to type {typeof(T).FullName}");
-    }
-}
-
-public class TestWebApplicationFactory<TProgram>
-    : WebApplicationFactory<TProgram> where TProgram : class {
-    public readonly static string ConnectionString = "Data Source=TestDb.db";
-
-    protected override void ConfigureWebHost(IWebHostBuilder builder) {
-        builder.ConfigureServices(services => {
-            var dbContextDescriptor = services.SingleOrDefault(
-                d => d.ServiceType == typeof(DbContextOptions<ApplicationDbContext>));
-            if (dbContextDescriptor is not null)
-                services.Remove(dbContextDescriptor);
-
-            var dbConnectionDescriptor = services.SingleOrDefault(
-                d => d.ServiceType == typeof(DbConnection));
-            if (dbConnectionDescriptor is not null)
-                services.Remove(dbConnectionDescriptor);
-
-            services.AddDbContext<ApplicationDbContext>(options => {
-                options.UseSqlite(ConnectionString);
-            });
-
-            var mailService = services.SingleOrDefault(
-                d => d.ServiceType == typeof(IConfirmationMailService));
-            if (mailService is not null)
-                services.Remove(mailService);
-
-            services.AddScoped<IConfirmationMailService, TestMailService>();
-
-            MigrateDbContext<ApplicationDbContext>(services);
-        });
-    }
-
-    public void MigrateDbContext<TContext>(IServiceCollection serviceCollection) where TContext : DbContext {
-        var serviceProvider = serviceCollection.BuildServiceProvider();
-
-        using var scope = serviceProvider.CreateScope();
-
-        var services = scope.ServiceProvider;
-        var context = services.GetService<TContext>();
-
-        if (context?.Database.IsNpgsql() ?? true) {
-            throw new Exception("Use Sqlite instead of sql server!");
-        }
-
-        context.Database.EnsureDeleted();
-
-        context.Database.Migrate();
-
     }
 }
