+++ InnoShop.UserManagerAPI/InnoShop.Infrastructure.UserManagerAPI.csproj
@@ -21,7 +21,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\InnoShop.Domain\InnoShop.Domain.csproj" />
-    <ProjectReference Include="..\InnoShop.Application.Shared\InnoShop.Application.Shared.csproj" />
+    <ProjectReference Include="..\InnoShop.Application\InnoShop.Application.csproj" />
   </ItemGroup>
 
 </Project>
